refactor(display): extract formatters into deep module, reduce display.js surface
Candidate 1 from architecture review: display.js was 824 lines with 22 exports
— formatters and renderers all public. Extracted pure formatting functions
into src/formatters.js. display.js now imports from formatters and re-exports
for backward compatibility. The formatting implementation is now internal to
the formatters module; display.js's own interface is the 5 render entry points
(displayCurrentWeather, display5DayForecast, display24HourForecast,
displayAlerts, displayMinutelyForecast) plus re-exports for existing consumers.

All 629 tests pass. Lint clean.
